requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.23.2
langchain==0.0.335
langchain-openai==0.0.2
openai==1.3.5
browseruse==0.0.1
playwright==1.40.0
pydantic==2.4.2
python-dotenv==1.0.0
httpx==0.25.1
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1 